requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
redis==5.0.1
sentence-transformers==2.2.2
pydantic==2.5.3
python-dotenv==1.0.0
pybind11>=2.12
huggingface_hub==0.14.1
